+++ packages/core/src/components/tooltip/tooltip.interfaces.ts
@@ -1,0 +1,3 @@
+export interface TooltipCloseEvent {
+  originalEvent?: MouseEvent | Event;
+}