+++ packages/dso-toolkit/src/components/tooltip/tooltip.models.ts
@@ -2,11 +2,19 @@ export const tooltipPositions = ["top", "right", "bottom", "left"] as const;
 
 export const tooltipStrategy = ["auto", "absolute", "fixed"] as const;
 
-export interface Tooltip {
+export interface Tooltip<TemplateFnReturnType> {
   active?: boolean;
   descriptive?: boolean;
   position: (typeof tooltipPositions)[number];
   strategy?: (typeof tooltipStrategy)[number];
-  label: string;
+  label?: string;
   id?: string;
+  heading?: TemplateFnReturnType;
+  content?: TemplateFnReturnType;
+  variant?: "onboarding";
+  dsoClose?: (e: CustomEvent<TooltipCloseEvent>) => void;
+}
+
+export interface TooltipCloseEvent {
+  originalEvent?: MouseEvent | Event;
 }