+++ packages/react/src/components/tooltip/tooltip.stories.tsx
@@ -1,9 +1,10 @@
-import type { Meta } from "@storybook/react";
-import { TooltipArgs, tooltipMeta, tooltipStories } from "dso-toolkit";
-
-import { HandlerFunction } from "@storybook/addon-actions";
 import * as React from "react";
+import type { Meta } from "@storybook/react";
+import { HandlerFunction } from "@storybook/addon-actions";
+
 import { templateContainer } from "../../templates";
+
+import { TooltipArgs, tooltipMeta, tooltipStories } from "dso-toolkit";
 
 import readme from "./readme.md?raw";
 
@@ -34,7 +35,29 @@ function asSiblingTemplate(tooltip: JSX.Element, id: string, action: HandlerFunc
   );
 }
 
-const { AsChild, AsSibling } = tooltipStories({
+const richContent = (
+  <>
+    <p>
+      Uitleg over wat u hier kunt doen en zit. Met een link naar{" "}
+      <a href="#" className="extern">
+        meer informatie
+      </a>
+      .
+    </p>
+    <div className="dso-button-row dso-align-right">
+      <button type="button" className="dso-primary">
+        <span>Oke!</span>
+        <svg className="di di-chevron-right">
+          <use href="dso-toolkit/dist/dso-icons.svg#chevron-right"></use>
+        </svg>
+      </button>
+    </div>
+  </>
+);
+
+const headingContent = <h5 slot="heading">Tip: Onboarding</h5>;
+
+const { AsChild, AsSibling, Onboarding } = tooltipStories({
   templateContainer,
   storyTemplates: (templates) => {
     const { tooltipTemplate } = templates;
@@ -43,8 +66,10 @@ const { AsChild, AsSibling } = tooltipStories({
       tooltipTemplate,
       asChildTemplate,
       asSiblingTemplate,
+      richContent,
+      headingContent,
     };
   },
 });
 
-export { AsChild, AsSibling };
+export { AsChild, AsSibling, Onboarding };