+++ storybook/src/components/tooltip/tooltip.core-stories.ts
@@ -1,7 +1,7 @@
 import type { Meta } from "@storybook/web-components";
 import { TooltipArgs, tooltipMeta, tooltipStories } from "dso-toolkit";
 
-import { asChildTemplate, asSiblingTemplate } from "./tooltip.content";
+import { asChildTemplate, asSiblingTemplate, headingContent, richContent } from "./tooltip.content";
 import { templateContainer } from "../../templates";
 
 import readme from "@dso-toolkit/core/src/components/tooltip/readme.md?raw";
@@ -13,7 +13,7 @@ const meta: Meta<TooltipArgs> = {
 
 export default meta;
 
-const { AsChild, AsSibling } = tooltipStories({
+const { AsChild, AsSibling, Onboarding } = tooltipStories({
   templateContainer,
   storyTemplates: (templates) => {
     const { tooltipTemplate } = templates;
@@ -22,8 +22,10 @@ const { AsChild, AsSibling } = tooltipStories({
       tooltipTemplate,
       asChildTemplate,
       asSiblingTemplate,
+      richContent: richContent(templates),
+      headingContent: headingContent(templates),
     };
   },
 });
 
-export { AsChild, AsSibling };
+export { AsChild, AsSibling, Onboarding };