+++ storybook/src/components/tooltip/tooltip.core-template.ts
@@ -1,20 +1,22 @@
 import { Tooltip } from "dso-toolkit";
-import { html } from "lit-html";
+import { html, TemplateResult } from "lit-html";
 import { ifDefined } from "lit-html/directives/if-defined.js";
 import { ComponentImplementation } from "../../templates";
 
-export const coreTooltip: ComponentImplementation<Tooltip> = {
+export const coreTooltip: ComponentImplementation<Tooltip<TemplateResult>> = {
   component: "tooltip",
   implementation: "core",
   template: () =>
-    function tooltipTemplate({ active, descriptive, position, label, id }) {
+    function tooltipTemplate({ active, descriptive, position, label, id, variant, dsoClose, content, heading }) {
       return html`<dso-tooltip
         ?descriptive=${ifDefined(descriptive)}
         position=${position}
         id=${ifDefined(id)}
         ?active=${active}
+        variant=${ifDefined(variant)}
+        @dsoClose=${dsoClose}
       >
-        ${label}
+        ${heading} ${label || content}
       </dso-tooltip>`;
     },
 };