+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Added
+* Tooltip: Nieuwe variant `onboarding` ([#3069](https://github.com/dso-toolkit/dso-toolkit/issues/3069))
+
 ### Task
 * **BREAKING** React: Update peerDependencies naar 19, verwijder 17 ([#3077](https://github.com/dso-toolkit/dso-toolkit/issues/3077))